requirements.txt
asgiref==3.6.0
Django==4.1.5
django-filter==22.1
djongo==1.3.6
dnspython==2.3.0
et-xmlfile==1.1.0
openpyxl==3.0.10
pdfkit==1.0.0
pymongo==3.12.3
pytz==2022.7.1
sqlparse==0.2.4
tzdata==2022.7
wkhtmltopdf==0.2
